# openapi.yaml
openapi: 3.0.3
info:
  title: Kibooz API
  version: "1"
  description: |
    REST API of Kibooz, the kindergarten app where teachers record the mood of children and parents follow it and
    earn points for sorting trash.

    Every JSON response uses one envelope. A success has `success: true` with an optional `message` and `data`.
    A failure has `success: false`, a `message` and a stable `errorCode`, and validation failures also carry a
    `details` object with one message per field.

    Protected routes need the access token from login as a bearer token. Rate limits answer `429` with `RATE_LIMITED`.
    They count requests per signed-in user, and login, registration and token routes per account, never per IP address.
  license:
    name: GNU General Public License v3.0
    url: https://github.com/SyafaHadyan/kibooz-backend/blob/main/LICENSE
servers:
  - url: https://kibooz-api.syafahadyan.com/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local development with the default port
...
security:
  - bearerAuth: []
paths:
...
  /classes/{classId}/videos:
    get:
      tags: [Classroom]
      operationId: listVideos
      summary: List the learning videos of a class
      description: Newest first. Open to the teachers of the class and the parents of a child in it.
      parameters:
        - $ref: "#/components/parameters/ClassPath"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: The videos of the requested page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoListEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
    post:
      tags: [Classroom]
      operationId: addVideo
      summary: Add a learning video to a class
      description: >-
        Only a teacher of the class can add one. The video is an https address. For a file the teacher recorded, first ask for a signed
        address with `POST /classes/{classId}/videos/upload-url`, send the file there, then add the video with the `videoUrl` of that
        answer. The file must have arrived by then, it must belong to this class, and it must be an accepted video.
      parameters:
        - $ref: "#/components/parameters/ClassPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddVideoRequest"
      responses:
        "201":
          description: Learning video added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        "502":
          $ref: "#/components/responses/StorageFailed"
  /classes/{classId}/videos/upload-url:
    post:
      tags: [Classroom]
      operationId: createVideoUpload
      summary: Get a signed address to upload a video file to
      description: >-
        Only a teacher of the class can ask for one. Send the file with a PUT request to `uploadUrl` before `expiresAt`, with exactly
        `sizeBytes` bytes and the headers of the answer, and no other header is needed. The bucket refuses a file of another size or
        type. Then add the video with `POST /classes/{classId}/videos` and the `videoUrl` of this answer. A file that is never added
        stays in the bucket until a lifecycle rule removes it.
      parameters:
        - $ref: "#/components/parameters/ClassPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoUploadRequest"
      responses:
        "200":
          description: Upload address created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoUploadEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        "502":
          $ref: "#/components/responses/StorageFailed"
        "503":
          description: File storage is not configured (`STORAGE_UNAVAILABLE`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Failure"
  /classes/{classId}/videos/{videoId}:
    patch:
      tags: [Classroom]
      operationId: updateVideo
      summary: Change the details of a learning video
      description: >-
        Any teacher of the class can change it. A missing field stays as it is, an empty `description` or `thumbnailUrl` clears it and
        a `durationSeconds` of 0 clears the duration. At least one field is required. The address of the video cannot change, so a
        different file is a new video.
      parameters:
        - $ref: "#/components/parameters/ClassPath"
        - $ref: "#/components/parameters/VideoPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVideoRequest"
      responses:
        "200":
          description: Learning video updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
    delete:
      tags: [Classroom]
      operationId: deleteVideo
      summary: Delete a learning video
      description: >-
        Any teacher of the class can delete it. When the video was uploaded and no other video uses the same file, the file is removed
        from storage too. A link to another site is only forgotten.
      parameters:
        - $ref: "#/components/parameters/ClassPath"
        - $ref: "#/components/parameters/VideoPath"
      responses:
        "200":
          description: Learning video deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /classes/{classId}/students:
    get:
      tags: [Classroom]
      operationId: listClassStudents
      summary: List the children of a class
      description: Only for the teachers of the class. Sorted by name, with the points, the rank and the latest mood of today.
      parameters:
        - $ref: "#/components/parameters/ClassPath"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: The children of the requested page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassStudentListEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
  /classes/{classId}/forum:
    get:
      tags: [Classroom]
      operationId: listThreads
      summary: List the forum threads of a class
      description: Newest thread first. Open to the teachers of the class and the parents of a child in it.
      parameters:
        - $ref: "#/components/parameters/ClassPath"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: The threads of the requested page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForumThreadListEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
    post:
      tags: [Classroom]
      operationId: createThread
      summary: Start a forum thread
      description: Open to the teachers of the class and the parents of a child in it.
      parameters:
        - $ref: "#/components/parameters/ClassPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateThreadRequest"
      responses:
        "201":
          description: Forum thread created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForumThreadEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
  /classes/{classId}/forum/{postId}/replies:
    get:
      tags: [Classroom]
      operationId: listReplies
      summary: List the replies to a forum thread
      description: Oldest reply first. `postId` is the id of a thread of the class.
      parameters:
        - $ref: "#/components/parameters/ClassPath"
        - $ref: "#/components/parameters/PostPath"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: The replies of the requested page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForumReplyListEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
    post:
      tags: [Classroom]
      operationId: createReply
      summary: Reply to a forum thread
      description: Open to the teachers of the class and the parents of a child in it. A reply cannot be answered again.
      parameters:
        - $ref: "#/components/parameters/ClassPath"
        - $ref: "#/components/parameters/PostPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReplyRequest"
      responses:
        "201":
          description: Forum reply added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForumReplyEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: The `token` returned by login or registration
  parameters:
...
    ClassPath:
      name: classId
      in: path
      required: true
      description: A class of the signed in user
      schema:
        type: string
        format: uuid
    VideoPath:
      name: videoId
      in: path
      required: true
      description: A learning video of the class
      schema:
        type: string
        format: uuid
    PostPath:
      name: postId
      in: path
      required: true
      description: A forum thread of the class
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      description: The page to return, starting at 1
      schema:
        type: integer
        minimum: 1
        maximum: 10000
        default: 1
    Limit:
      name: limit
      in: query
      description: How many items one page holds
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
  responses:
    ValidationError:
      description: The input is invalid (`VALIDATION_ERROR`), with one message per field in `details`
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Failure"
          example:
            success: false
            message: The submitted data is invalid
            errorCode: VALIDATION_ERROR
            details:
              email: must be a valid email address
    Unauthorized:
      description: The access token is missing, invalid or expired (`AUTH_TOKEN_MISSING`, `AUTH_TOKEN_INVALID`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Failure"
    Forbidden:
      description: The role or the data is not allowed for this user (`AUTH_FORBIDDEN`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Failure"
    NotFound:
      description: |
        The resource does not exist or is not visible to this user (`CLASS_NOT_FOUND`, `CLASS_CODE_NOT_FOUND`,
        `STUDENT_NOT_FOUND`, `PROFILE_NOT_FOUND`, `GUIDANCE_NOT_FOUND`, `FORUM_POST_NOT_FOUND`, `VIDEO_NOT_FOUND`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Failure"
    RateLimited:
      description: Too many requests from this user or for this account (`RATE_LIMITED`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Failure"
...
    StorageFailed:
      description: The file could not be stored (`STORAGE_ERROR`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Failure"
  schemas:
    Role:
      type: string
      enum: [GURU, WALI, ADMIN]
      description: Teacher, parent or administrator. ADMIN cannot be registered.
...
    Failure:
      type: object
      required: [success, message, errorCode]
      properties:
        success:
          type: boolean
          enum: [false]
        message:
          type: string
        errorCode:
          type: string
          description: A stable code for the client to branch on
        details:
          type: object
          description: One message per invalid field
          additionalProperties:
            type: string
    EmptyEnvelope:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          enum: [true]
        message:
          type: string
...
    PageInfo:
      type: object
      required: [page, limit, total]
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
          maximum: 50
        total:
          type: integer
          minimum: 0
          description: How many items the whole list holds
    AddVideoRequest:
      type: object
      required: [title, videoUrl]
      properties:
        title:
          type: string
          maxLength: 150
        description:
          type: string
          maxLength: 2000
        videoUrl:
          type: string
          format: uri
          maxLength: 2048
          description: An https address the app can stream
        thumbnailUrl:
          type: string
          format: uri
          maxLength: 2048
          description: An https address of a preview image
        durationSeconds:
          type: integer
          minimum: 1
          maximum: 86400
    UpdateVideoRequest:
      type: object
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 150
        description:
          type: string
          maxLength: 2000
          description: An empty text clears the description
        thumbnailUrl:
          type: string
          maxLength: 2048
          description: An https address of a preview image, an empty text clears it
        durationSeconds:
          type: integer
          minimum: 0
          maximum: 86400
          description: 0 clears the duration
    VideoUploadRequest:
      type: object
      required: [contentType, sizeBytes]
      properties:
        contentType:
          type: string
          enum: [video/mp4, video/webm]
        sizeBytes:
          type: integer
          format: int64
          minimum: 1
          description: The exact size of the file, the server refuses a larger file than `VIDEO_MAX_MB`
    VideoUpload:
      type: object
      required: [uploadUrl, method, headers, videoUrl, expiresAt]
      properties:
        uploadUrl:
          type: string
          format: uri
          description: The signed address to send the file to
        method:
          type: string
          enum: [PUT]
        headers:
          type: object
          description: Headers the upload request must carry
          required: [Content-Type]
          properties:
            Content-Type:
              type: string
        videoUrl:
          type: string
          format: uri
          description: The public address of the file, to add as the video once the upload is done
        expiresAt:
          type: string
          format: date-time
    VideoUploadEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        message:
          type: string
        data:
          $ref: "#/components/schemas/VideoUpload"
    Video:
      type: object
      required: [id, title, description, videoUrl, thumbnailUrl, durationSeconds, createdAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        videoUrl:
          type: string
          format: uri
        thumbnailUrl:
          type: string
          format: uri
          nullable: true
        durationSeconds:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
    VideoList:
      allOf:
        - $ref: "#/components/schemas/PageInfo"
        - type: object
          required: [videos]
          properties:
            videos:
              type: array
              items:
                $ref: "#/components/schemas/Video"
    VideoEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        message:
          type: string
        data:
          $ref: "#/components/schemas/Video"
    VideoListEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        message:
          type: string
        data:
          $ref: "#/components/schemas/VideoList"
    ForumAuthor:
      type: object
      required: [id, fullName, role, avatarUrl]
      description: The account that wrote a post. A deleted account shows the name `Deleted account` and no photo.
      properties:
        id:
          type: string
          format: uuid
          description: The id of the account, compare it with the id of the signed in user
        fullName:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        avatarUrl:
          type: string
          format: uri
          nullable: true
    CreateThreadRequest:
      type: object
      required: [title, body]
      properties:
        title:
          type: string
          maxLength: 150
        body:
          type: string
          maxLength: 5000
    ForumThread:
      type: object
      required: [id, title, body, author, replyCount, createdAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        body:
          type: string
        author:
          $ref: "#/components/schemas/ForumAuthor"
        replyCount:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
    ForumThreadList:
      allOf:
        - $ref: "#/components/schemas/PageInfo"
        - type: object
          required: [threads]
          properties:
            threads:
              type: array
              items:
                $ref: "#/components/schemas/ForumThread"
    ForumThreadEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        message:
          type: string
        data:
          $ref: "#/components/schemas/ForumThread"
    ForumThreadListEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        message:
          type: string
        data:
          $ref: "#/components/schemas/ForumThreadList"
    CreateReplyRequest:
      type: object
      required: [body]
      properties:
        body:
          type: string
          maxLength: 5000
    ForumReply:
      type: object
      required: [id, body, author, createdAt]
      properties:
        id:
          type: string
          format: uuid
        body:
          type: string
        author:
          $ref: "#/components/schemas/ForumAuthor"
        createdAt:
          type: string
          format: date-time
    ForumReplyList:
      allOf:
        - $ref: "#/components/schemas/PageInfo"
        - type: object
          required: [replies]
          properties:
            replies:
              type: array
              items:
                $ref: "#/components/schemas/ForumReply"
    ForumReplyEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        message:
          type: string
        data:
          $ref: "#/components/schemas/ForumReply"
    ForumReplyListEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        message:
          type: string
        data:
          $ref: "#/components/schemas/ForumReplyList"
...
    ClassStudent:
      type: object
      required: [id, fullName, nisn, avatarUrl, currentPoints, classRank, todayMood]
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        nisn:
          type: string
        avatarUrl:
          type: string
          format: uri
          nullable: true
        currentPoints:
          type: integer
          minimum: 0
        classRank:
          type: integer
          minimum: 0
        todayMood:
          type: string
          enum: [SENANG, SEDIH, MARAH, BINGUNG, null]
          nullable: true
          description: The latest mood recorded today, null when none was
    ClassStudentList:
      allOf:
        - $ref: "#/components/schemas/PageInfo"
        - type: object
          required: [students]
          properties:
            students:
              type: array
              items:
                $ref: "#/components/schemas/ClassStudent"
    ClassStudentListEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        message:
          type: string
        data:
          $ref: "#/components/schemas/ClassStudentList"
...